+++ backend/src/main/java/com/spendingstracker/app/dto/response/SpendingResponse.java
@@ -4,14 +4,13 @@ import lombok.Builder;
 import lombok.Data;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
-/**
- * Object returned to client with relevant spending information
- */
+/** Object returned to client with relevant spending information */
 @Builder
 @Data
 public class SpendingResponse {
-    private long spendingId;
+    private BigInteger spendingId;
     private String category;
     private BigDecimal amount;
 }
