+++ backend/src/main/java/com/spendingstracker/app/repository/UserRepository.java
@@ -5,9 +5,11 @@ import com.spendingstracker.app.entity.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 /** <code>JpaRepository</code> used to make queries to the <code>APP>.USER</code> table. */
 @Repository
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, BigInteger> {
     /**
      * Find <code>User</code> by <code>username</code>
      *
